Cambiata directory e rimosse funzionalità inutili
