--- esempio_AIP/logicProblem.py
@@ -1,111 +1,0 @@
-# logicProblem.py - Representations Logics
-# AIFCA Python3 code Version 0.8.0 Documentation at http://aipython.org
-
-# Artificial Intelligence: Foundations of Computational Agents
-# http://artint.info
-# Copyright David L Poole and Alan K Mackworth 2017.
-# This work is licensed under a Creative Commons
-# Attribution-NonCommercial-ShareAlike 4.0 International License.
-# See: http://creativecommons.org/licenses/by-nc-sa/4.0/deed.en
-
-class Clause(object):
-    """A definite clause"""
-
-    def __init__(self,head,body=[]):
-        """clause with atom head and lost of atoms body"""
-        self.head=head
-        self.body = body
-
-    def __str__(self):
-        """returns the string representation of a clause.
-        """
-        if self.body:
-            return self.head + " <- " + " & ".join(self.body) + "."
-        else:
-            return self.head + "."
-
-    def isAtom(self):
-        return not self.body
-
-class Askable(object):
-    """An askable atom"""
-
-    def __init__(self,atom):
-        """clause with atom head and lost of atoms body"""
-        self.atom=atom
-
-    def __str__(self):
-        """returns the string representation of a clause."""
-        return "askable " + self.atom + "."
-
-def yes(ans):
-    """returns true if the answer is yes in some form"""
-    return ans.lower() in ['yes', 'yes.', 'oui', 'oui.', 'y', 'y.']    # bilingual
-
-#from display import Displayable
-        
-class KB():#Displayable
-    """A knowledge base consists of a set of clauses.
-    This also creates a dictionary to give fast access to the clauses with an atom in head.
-    """
-    def __init__(self, statements=[]):
-        self.statements = statements
-        self.clauses = [c for c in statements if isinstance(c, Clause)]
-        self.askables = [c.atom for c in statements if isinstance(c, Askable)]
-        self.atom_to_clauses = {}  # dictionary giving clauses with atom as head
-        for c in self.clauses:
-            if c.head in self.atom_to_clauses:
-                self.atom_to_clauses[c.head].add(c)
-            else:
-                self.atom_to_clauses[c.head] = {c}
-
-    def clauses_for_atom(self,a):
-        """returns set of clauses with atom a as the head"""
-        if a in self.atom_to_clauses:
-            return  self.atom_to_clauses[a]
-        else:
-            return set()
-
-    def __str__(self):
-        """returns a string representation of this knowledge base.
-        """
-        return '\n'.join([str(c) for c in self.statements])
-
-triv_KB = KB([
-    Clause('i_am', ['i_think']),
-    Clause('i_think'),
-    Clause('i_smell', ['i_exist'])
-    ])
-
-elect = KB([
-    Clause('light_l1'),
-    Clause('light_l2'),
-    Clause('ok_l1'),
-    Clause('ok_l2'),
-    Clause('ok_cb1'),
-    Clause('ok_cb2'),
-    Clause('live_outside'),
-    Clause('live_l1', ['live_w0']),
-    Clause('live_w0', ['up_s2','live_w1']),
-    Clause('live_w0', ['down_s2','live_w2']),
-    Clause('live_w1', ['up_s1', 'live_w3']),
-    Clause('live_w2', ['down_s1','live_w3' ]),
-    Clause('live_l2', ['live_w4']),
-    Clause('live_w4', ['up_s3','live_w3' ]),
-    Clause('live_p_1', ['live_w3']),
-    Clause('live_w3', ['live_w5', 'ok_cb1']),
-    Clause('live_p_2', ['live_w6']),
-    Clause('live_w6', ['live_w5', 'ok_cb2']),
-    Clause('live_w5', ['live_outside']),
-    Clause('lit_l1', ['light_l1', 'live_l1', 'ok_l1']),
-    Clause('lit_l2', ['light_l2', 'live_l2', 'ok_l2']),
-    Askable('up_s1'),
-    Askable('down_s1'),
-    Askable('up_s2'),
-    Askable('down_s2'),
-    Askable('up_s3'),
-    Askable('down_s2')
-    ])
-
-
-#print(kb)
