+++ src/__pycache__/logicBottomUp.cpython-36.pyc
